add comment

+++ src/main/java/com/java/community/controller/PublishController.java
@@ -48,7 +48,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model){
 
@@ -89,7 +89,7 @@ public class PublishController {
      * @return
      */
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(value = "id") Integer id,
+    public String edit(@PathVariable(value = "id") Long id,
                         Model model){
         QuestionDTO question = questionService.findById(id);
         model.addAttribute("title",question.getTitle());
